requirements.txt
Flask==2.2.5
psycopg2-binary==2.9.7
youtube-transcript-api==0.6.0
openai==0.27.8
PyYAML==6.0.1